+++ ui/src/__tests__/App.spec.tsx
@@ -1,7 +1,7 @@
-import './installSesLockdown';
+import '../installSesLockdown';
 import { render, screen } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
-import App from './App';
+import App from '../App';
 
 describe('App.tsx', () => {
   it('renders app title', async () => {
